// packages/neko-types/src/types/asset/entity.d.ts
export type AssetMediaType = 'video' | 'audio' | 'image' | 'sequence' | 'text';